Update README.md
The farmers can sell their productions online and the buyer can purchase various agricultural 
products online. Buyer can send purchase request to check the quality of the product. After 
collecting all the farm produce from the farmers, it should be sold to the customers. This project
covers these entries and the data collections.There are 2 types of users: Customer & Farmers.The 
login id and password must be required to login the system. The article and agro products section 
helps farmers to share their products and increase profitability.